backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt==4.0.1
passlib
python-dotenv
email-validator
